Upgraded to export data
Creates an object which randomly selects images and augments with existing segmentation, placing the object at the median sidewalk location.  Item can be cycled with ".forward()" and ".previous()".  Item to augment with, the threshold (number of pixels) for selection can be updated with ".update_label()" and ".update_thresh()".  the augmented image and JSON can be saved with ".save()", or you can skip to the next random image with ".next()".  The class needs to be told the location of original images, JSON labels, and an output directory when initiating.  I was using unprocessed cityscape data separated into our train/test/val split.